Move tree crawler functions to tree
