+++ src/component.js
@@ -3,7 +3,8 @@ import { commitRoot } from './diff/commit';
 import options from './options';
 import { createVNode, Fragment } from './create-element';
 import { patch } from './diff/patch';
-import { TYPE_COMPONENT, DIRTY_BIT, FORCE_UPDATE } from './constants';
+import { DIRTY_BIT, FORCE_UPDATE } from './constants';
+import { getDomSibling, updateParentDomPointers } from './tree';
 
 /**
  * Base Component class. Provides `setState()` and `forceUpdate()`, which
@@ -84,42 +85,6 @@ Component.prototype.forceUpdate = function(callback) {
 Component.prototype.render = Fragment;
 
 /**
- * @param {import('./internal').Internal} internal
- * @param {number | null} [childIndex]
- * @returns {import('./internal').PreactNode}
- */
-export function getDomSibling(internal, childIndex) {
-	if (childIndex == null) {
-		// Use childIndex==null as a signal to resume the search from the vnode's sibling
-		return internal._parent
-			? getDomSibling(
-					internal._parent,
-					internal._parent._children.indexOf(internal) + 1
-			  )
-			: null;
-	}
-
-	let sibling;
-	for (; childIndex < internal._children.length; childIndex++) {
-		sibling = internal._children[childIndex];
-
-		if (sibling != null && sibling._dom != null) {
-			// Since updateParentDomPointers keeps _dom pointer correct,
-			// we can rely on _dom to tell us if this subtree contains a
-			// rendered DOM node, and what the first rendered DOM node is
-			return sibling._dom;
-		}
-	}
-
-	// If we get here, we have not found a DOM node in this vnode's children.
-	// We must resume from this vnode's sibling (in it's parent _children array)
-	// Only climb up and search the parent if we aren't searching through a DOM
-	// VNode (meaning we reached the DOM parent of the original vnode that began
-	// the search)
-	return internal._flags & TYPE_COMPONENT ? getDomSibling(internal) : null;
-}
-
-/**
  * Trigger in-place re-rendering of a component.
  * @param {import('./internal').Component} component The component to rerender
  */
@@ -152,24 +117,6 @@ function rerenderComponent(component) {
 		if (internal._dom != startDom) {
 			updateParentDomPointers(internal);
 		}
-	}
-}
-
-/**
- * @param {import('./internal').Internal} internal
- */
-function updateParentDomPointers(internal) {
-	if ((internal = internal._parent) != null && internal._component != null) {
-		internal._dom = null;
-		for (let i = 0; i < internal._children.length; i++) {
-			let child = internal._children[i];
-			if (child != null && child._dom != null) {
-				internal._dom = child._dom;
-				break;
-			}
-		}
-
-		return updateParentDomPointers(internal);
 	}
 }
